+++ R/general.R
@@ -84,16 +84,12 @@ general.default <- function(x, y, features="all", summary=c("mean", "sd"),
   if(nrow(x) != length(y)) {
     stop("x and y must have same number of rows")
   }
-  
-  if (nlevels(y) > length(y) / 10) {
-    stop("y must contain classes values")
-  }
 
   if(features[1] == "all") {
     features <- ls.general()
   }
   features <- match.arg(features, ls.general(), TRUE)
-  colnames(x) <- make.names(colnames(x))
+  colnames(x) <- make.names(colnames(x), unique=TRUE)
   
   if (length(summary) == 0) {
     summary <- "non.aggregated"